refactor: move MaterialDataAppliers to material data import folder
